src/main/webapp/app/entities/technology/route/technology-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { ITechnology, Technology } from '../technology.model';
import { TechnologyService } from '../service/technology.service';

import { TechnologyRoutingResolveService } from './technology-routing-resolve.service';

describe('Technology routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: TechnologyRoutingResolveService;
  let service: TechnologyService;
  let resultTechnology: ITechnology | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(TechnologyRoutingResolveService);
    service = TestBed.inject(TechnologyService);
    resultTechnology = undefined;
  });

  describe('resolve', () => {
    it('should return ITechnology returned by find', () => {
      // GIVEN
      service.find = jest.fn(technologyID => of(new HttpResponse({ body: { technologyID } })));
      mockActivatedRouteSnapshot.params = { technologyID: '9fec3727-3421-4967-b213-ba36557ca194' };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultTechnology = result;
      });

      // THEN
      expect(service.find).toBeCalledWith('9fec3727-3421-4967-b213-ba36557ca194');
      expect(resultTechnology).toEqual({ technologyID: '9fec3727-3421-4967-b213-ba36557ca194' });
    });

    it('should return new ITechnology if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultTechnology = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultTechnology).toEqual(new Technology());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as Technology })));
      mockActivatedRouteSnapshot.params = { technologyID: '9fec3727-3421-4967-b213-ba36557ca194' };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultTechnology = result;
      });

      // THEN
      expect(service.find).toBeCalledWith('9fec3727-3421-4967-b213-ba36557ca194');
      expect(resultTechnology).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
